+++ gdb/samples/signal.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+make hello
+gdb --quiet ./hello -x signal.gdb